chore: Add PR test workflow
This commit adds a new GitHub Actions workflow file for testing pull requests. The workflow is triggered on the oc_5019 branch and runs on Ubuntu latest. It currently includes a step to log a success message. Additional steps for setting up Node.js, installing dependencies, and installing the Ionic CLI are commented out but can be uncommented as needed.

chore: Update PR test workflow

- Set up Node.js version 18
- Install project dependencies
- Install Ionic CLI globally
- Create Cypress.env.json file with secrets
- Display content of Cypress.env.json file
- Start application and run tests using Cypress.io GitHub Action
- Upload Cypress screenshots on test failure

chore: Update upload-artifact action to v4 in PR test workflow

The commit updates the version of the upload-artifact action used in the PR test workflow from v3 to v4. This change ensures compatibility with the latest features and improvements provided by the updated version.

chore: Add step to checkout repository in PR test workflow

This commit adds a new step to the PR test workflow that checks out the repository before running the tests. This ensures that the latest code changes are used during testing.